Separate extension am dto
